Moved chunking logic from CLI to settings.json Setting added: verboseLogging & deleteUploadChunks. Removed unneeded CLI arguments

+++ main.py
@@ -24,9 +24,13 @@ def main():
 
 	# Get configurations & set variables
 	settings = utils.read_configuration_settings()
+	verbose_endpoint_logging = settings["verboseEndpointLogging"]
 	oauth_service_uri = settings["uris"]["oauthService"]
 	integration_api_uri = settings["uris"]["integrationApi"]
 	thread_count = settings["threadCount"]
+	compress_upload_chunks = settings["compressUploadChunks"]
+	upload_chunk_size_mb = settings["uploadChunkSizeMb"]
+	delete_upload_chunks = settings["deleteUploadChunks"]
 	database = settings["database"]
 	rotatable_token = settings["rotatableToken"]
 	workspace_id = settings["workspaceId"]
@@ -56,20 +60,19 @@ def main():
 	refresh_token = anaplan_oauth.refresh_token_thread(1, name="Refresh Token", delay=2000, uri=f'{oauth_service_uri}/token', database=database, rotatable_token=settings["rotatableToken"])
 	refresh_token.start()
 
-	# Set File to upload, chunk size, and compression toggle
+	# Set File to upload and import data source
 	file_to_upload = args.file_to_upload
-	chunk_size_mb = args.chunk_size_mb
 	import_data_source = args.import_data_source
-	no_compression = args.no_compression
-
+	
 	# Chunk files
-	chunk_files = file_ops.write_chunked_files(file=file_to_upload, chunk_size_mb=chunk_size_mb, no_compression=no_compression)
+	chunk_files = file_ops.write_chunked_files(file=file_to_upload, chunk_size_mb=upload_chunk_size_mb, compress_upload_chunks=compress_upload_chunks)
 
 	# Upload files to Anaplan
-	anaplan_ops.upload_all_chunks(file_to_upload=file_to_upload, import_data_source=import_data_source, chunk_files=chunk_files, no_compression=no_compression, max_workers=thread_count, base_uri=integration_api_uri, workspace_id=workspace_id, model_id=model_id)  
+	anaplan_ops.upload_all_chunks(file_to_upload=file_to_upload, import_data_source=import_data_source, chunk_files=chunk_files, compress_upload_chunks=compress_upload_chunks, max_workers=thread_count, verbose_endpoint_logging = verbose_endpoint_logging, base_uri=integration_api_uri, workspace_id=workspace_id, model_id=model_id)  
 
 	# Delete temporary files
-	file_ops.delete_files(chunk_files)
+	if delete_upload_chunks:
+		file_ops.delete_files(chunk_files)
 
 	print('Process complete. Exiting...')
 	logger.info('Process complete. Exiting...')
